\name{adaptedSampledSymmetricRandomWalkGenerator}
\alias{adaptedSampledSymmetricRandomWalkGenerator}
\title{Generator of multiple paths of sampled symmetric random walk along with the description of each of its steps}
\usage{
adaptedSampledSymmetricRandomWalkGenerator( steps = 3000,
                                            prob = c('head' = 0.5, 'tail' = 0.5),
                                            n = 1)
}
\arguments{
\item{steps}{Number of steps of the random walk}
\item{prob}{Probability of occurence of head and tail for each step}
\item{n}{Number of sample generated by the function}
}
\description{
Generates a list of symmetric random walk sampled n times along with its time periods.
The time period goes from 0 up to T (0: steps), for each sample.
It corresponds to a single random walk's path with each of its associated time period.
}
\examples{
# Generate a list of 20 sampled symmetric random walks
adaptedSampledSymmetricRandomWalkGenerator(n = 20)
}
