% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmotion.R
\name{get_sbmotion_point}
\alias{get_sbmotion_point}
\title{Brownian motion pointwise generator}
\usage{
get_sbmotion_point(time_to_maturity = 4, prob = c(head = 0.5, tail = 0.5),
  seed = 1)
}
\arguments{
\item{seed}{- Define the seed sequence to alway get the same sampled
randomly generated brownian motion.}
}
\description{
Brownian motion pointwise generator
}
