% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticVariation.R
\name{quadraticVariation}
\alias{quadraticVariation}
\title{Computation of the quadratic variation of random walk.
If random walk is symmetric and thus probability of obtain head equal the probabilty
of getting tail, the quadatric variation of this symmetric random walk
equal the variance of the random walk.}
\usage{
quadraticVariation(W, to)
}
\arguments{
\item{steps}{- Number of steps of the random walk}
}
\description{
Computation of the quadratic variation of random walk.
If random walk is symmetric and thus probability of obtain head equal the probabilty
of getting tail, the quadatric variation of this symmetric random walk
equal the variance of the random walk.
}
\examples{

}
\references{
[1] Symmetric Random Walk, Stochastic Calculus for Finance, Steven e. Shreve, 2004, pp 85
}
