% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwexp.R
\name{rwexp}
\alias{rwexp}
\title{Expectation of random walk.
The expectation of the random walk as described in "stochastic calculus
for finance" equal 0.
Therefore the calculation made by this function is only a verification
of the Expected zero growth for random walk.
More details and calculation will come later.}
\usage{
rwexp(Mk)
}
\description{
Expectation of random walk.
The expectation of the random walk as described in "stochastic calculus
for finance" equal 0.
Therefore the calculation made by this function is only a verification
of the Expected zero growth for random walk.
More details and calculation will come later.
}
