\name{randomWalkGenerator}
\alias{randomWalkGenerator}
\title{Generator of random walk}
\usage{
randomWalkGenerator(steps = 3000,
                    prob = c('head' = 0.5, 'tail' = 0.5),
                    n = 1)
}
\arguments{
\item{steps}{Number of steps of the random walk}
\item{prob}{Probability of occurence of head and tails for each step}
\item{n}{Number of random walk generated by the function}
}
\description{
Generate a symmetric random walk with step size one.
By default the probability of getting head or tail in one coin toss is one half.
}
\examples{
randomWalkGenerator()
}
