% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srwalkGenerator.R
\name{srwalkGenerator}
\alias{srwalkGenerator}
\title{It generates a list of n Sampled Random Walks along with the appropriate time period.
The time period goes from 0 up to T (0: time_to_maturity).
Because it is the same experiment for all the items of the list,
each of these depends on the same time period.}
\usage{
srwalkGenerator(time_to_maturity = 100, prob = c(head = 0.5, tail = 0.5),
  scale = 1, seed = 1, n = 1)
}
\arguments{
\item{time_to_maturity}{- Number of time_to_maturity of the random walk}

\item{prob}{- Probability of occurence of head and tail for each step}

\item{scale}{- It defines the time partition between each unit of time.
For instance if the scale is 100, the time step [0,1] will be cut in 100 parts.
Furthermore this param defines the increment multiplier as well.
If the scale is 100, the multiplier which defines the next value will be
(+1 / -1) * 1/sqrt(100)}

\item{seed}{- With same seed, 2 exeriments will give the same output. It therefore
provides reproducibles experiments.}

\item{n}{- Number of samples generated by the function}
}
\description{
It generates a list of n Sampled Random Walks along with the appropriate time period.
The time period goes from 0 up to T (0: time_to_maturity).
Because it is the same experiment for all the items of the list,
each of these depends on the same time period.
}
\examples{
# Generate a list of 20 sampled symmetric random walks
srwalkGenerator(n = 20)
}
\references{
[1] Symmetric Random Walk, Stochastic Calculus for Finance, Steven e. Shreve, 2004, pp 83-84
}
