% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trwalkGenerator.R
\name{trwalkGenerator}
\alias{trwalkGenerator}
\title{trwalkGenerator compute all the sets of the values taken by a random walk
along with the associated probabilities.}
\usage{
trwalkGenerator(time_to_maturity = 100, prob = c(head = 1/2, tail = 1/2),
  scale = 1, full = FALSE)
}
\arguments{
\item{time_to_maturity}{- Time at which the random walk process stops}

\item{prob}{- Probability of occurence of head and tail for each step}

\item{scale}{- Divider of the time step and multiplier of the increments}
}
\description{
trwalkGenerator compute all the sets of the values taken by a random walk
along with the associated probabilities.
}
\examples{
# Generate the theoretical distribution and value for a symmetric random walk
process up to time_to_maturity = 4 and scale = 100:
trwalkGenerator(time_to_maturity = 4, scale = 100)
}
\references{
[1] Symmetric Random Walk, Stochastic Calculus for Finance, Steven e. Shreve, 2004, pp 83-84
}
