\name{sampledSymmetricRandomWalkGenerator}
\alias{sampledSymmetricRandomWalkGenerator}
\title{Generator of multiple paths of sampled symmetric random walk}
\usage{
sampledSymmetricRandomWalkGenerator(  steps = 3000,
                                      prob = c('head' = 0.5, 'tail' = 0.5),
                                      n = 1)
}
\arguments{
\item{steps}{Number of steps of the random walk}
\item{prob}{Probability of occurence of head and tail for each step}
\item{n}{Number of sample generated by the function}
}
\description{
Generator of symmetric random walk sampled n times.
}
\examples{
# Generate 20 samples of a symmetric random walk
sampledSymmetricRandomWalkGenerator(n = 20)
}
