% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmotion.R
\name{sbmotion}
\alias{sbmotion}
\title{Brownian motion generator}
\usage{
sbmotion(time_to_maturity = 4, seed = 1, scale = 100)
}
\arguments{
\item{seed}{- Define the seed sequence to alway get the same sampled
randomly generated brownian motion.}
}
\description{
Brownian motion generator
}
