% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmotionGenerator.R
\name{sbmotionGenerator}
\alias{sbmotionGenerator}
\title{A generator of Sampled Brownian Motion}
\usage{
sbmotionGenerator(time_to_maturity = 4, seed = 1, scale = 100, n = 1)
}
\arguments{
\item{time_to_maturity}{Maximum time up to the Brownian Motion evolves}

\item{seed}{With same seed, 2 exeriments will(would, in fact not yet the
case when the param scale is different) give the same output. It therefore
provides reproducibles experiments.}

\item{scale}{It defines the time partition between each unit of time. For
instance if the scale is 100, the time step [0,1] will be cut in 100 parts.}

\item{n}{Number of samples generated by the function}
}
\value{
\code{sbmotionGenerator()} outputs a \bold{list of data.frame}.
Each data.frame in the list contains the whole range of time period and
the value of the corresponding Brownian Motion at that time.
}
\description{
It generates multiple (n) Sampled Brownian Motions along with
  the appropriate time period. The time period goes from 0 up to T (0:
  time_to_maturity).
}
\examples{
# Generate a list of 20 sampled Brownian Motions
sbmotionGenerator(n = 20)


}
\references{
[1] Brownian Motion, Stochastic Calculus for Finance,
  Steven e. Shreve, 2004, pp 93-97
}
