% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmotionGenerator.R
\name{sbmotionGenerator}
\alias{sbmotionGenerator}
\title{Brownian motion generator}
\usage{
sbmotionGenerator(time_to_maturity = 4, prob = c(head = 1/2, tail = 1/2),
  seed = 1, scale = 100, n = 1, reproducible = F)
}
\arguments{
\item{seed}{- Define the seed sequence to alway get the same sampled
randomly generated brownian motion.}
}
\description{
Brownian motion generator
}
